moved abbreviation method
